Divide .cpp e .hpp
-Classes CamCap, KalmanFilter, Strategy e StrategyGUI foram divididas em arquivos .hpp e .cpp, com novos arquivos inclusos no CMakeLists.txt
-Correção do bug que impedia inclusão de LS.cpp e CPUTimer.cpp no CMakeLists.txt
-Makefile apagado por não ser mais usado e nem atualizado
-Pasta do CLion .idea/codeStyles/ incluída no git para padronizar o estilo de código em todos os PCs
